refactor: update acceptance test
- use MainWindowMenuHandler for the test
