+++ test-acceptance/src/org/omegat/gui/main/TestCoreGUI.java
@@ -69,7 +69,7 @@ public abstract class TestCoreGUI extends AssertJSwingJUnitTestCase {
         frame = GuiActionRunner.execute(() -> {
             Core.setProject(new NotLoadedProject());
             UIDesignManager.initialize();
-            TestMainWindow mw = new TestMainWindow(TestMainWindowMenuHandler.class);
+            TestMainWindow mw = new TestMainWindow(new MainWindowMenuHandler());
             TestCoreInitializer.initMainWindow(mw);
             TestCoreInitializer.initAutoSave(autoSave);
 
@@ -94,22 +94,4 @@ public abstract class TestCoreGUI extends AssertJSwingJUnitTestCase {
         }
     };
 
-    static class TestMainWindowMenu extends BaseMainWindowMenu {
-
-        TestMainWindowMenu(IMainWindow mainWindow, BaseMainWindowMenuHandler mainWindowMenuHandler) {
-            super(mainWindow, mainWindowMenuHandler);
-            initComponents();
-        }
-
-        @Override
-        void createMenuBar() {
-            mainMenu.add(projectMenu);
-            mainMenu.add(editMenu);
-            mainMenu.add(gotoMenu);
-            mainMenu.add(viewMenu);
-            mainMenu.add(toolsMenu);
-            mainMenu.add(optionsMenu);
-            mainMenu.add(helpMenu);
-        }
-    }
 }
